disabled organizer registration, as it has same view as participant

+++ src/app/services/account/auth.service.ts
@@ -147,7 +147,7 @@ export class AuthService {
     if (this.platform.is('capacitor')) {
       await Toast.show({
         text: message,
-        duration: 'short',
+        duration: 'long',
         position: 'bottom',
       });
     } else {
